# api/openapi.yaml
openapi: 3.0.0
info:
  title: Online Cinema Service
  description: An online cinema service that allows admins to add halls, movies, and cinema sessions, and allows users to view the schedule, see movies watched, and buy tickets. The service returns tickets as PDF documents and exposes a RESTful API.
  version: 1.0.0
servers:
  - url: http://localhost
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Incorrect request was sent to the server.
    Unauthorized:
      description: Unauthorized request, user not authenticated.
    Forbidden:
      description: The user does not have permission to access the requested resource.
    NotFound:
      description: The requested resource was not found on the server.
    InternalServerError:
      description: An unexpected error occurred on the server.
  schemas:
...
    CinemaSession:
      type: object
      properties:
        id:
          type: integer
          example: 1
          description: Unique identifier of the cinema session. Generated on the server side.
          readOnly: true
        movie_id:
          type: integer
          example: 1
          description: ID of the movie being shown in the cinema session.
        hall_id:
          type: integer
          example: 1
          description: ID of the hall where the session will take place.
        start_time:
          type: string
          format: timestamp
          description: Start time of the cinema session
          example: 2024-05-18 20:00:00
        end_time:
          type: string
          format: timestamp
          description: Start time of the cinema session.
          example: 2024-05-18 22:00:00
          readOnly: true
        price:
          type: number
          format: float
          example: 10.5
          description: Price for the cinema session in GEL
        status:
          type: string
          description: Current status of the movie
          example: scheduled
          readOnly: true
    CinemaSessionResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
          description: Unique identifier of the cinema session. Generated on the server side.
          readOnly: true
        movie_id:
          type: integer
          example: 1
          description: ID of the movie being shown in the cinema session.
        hall_id:
          type: integer
          example: 1
          description: ID of the hall where the session will take place.
        start_time:
          type: string
          format: timestamp
          description: Start time of the cinema session
          example: 2024-05-18 20:00:00
        price:
          type: number
          format: float
          example: 10.5
          description: Price for the cinema session in GEL
        status:
          type: string
          description: Current status of the movie
          example: scheduled
          readOnly: true
...
paths:
...
  /cinema-sessions:
    get:
      tags:
        - cinema sessions
      summary: Returns a list of all cinema sessions for all halls
      operationId: getAllCinemaSessions
      parameters:
        - name: offset
          in: query
          description: The number of records to be skipped
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: The number of records to be returned
          required: false
          schema:
            type: integer
        - name: date
          in: query
          description: The date for which the movie sessions should be displayed in format YYYY-MM-DD
          required: false
          schema:
            type: string
            format: date

      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CinemaSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []
  /cinema-sessions/{hallId}:
    get:
      tags:
        - cinema sessions
      summary: Returns all cinema sessions
      operationId: getCinemaSessionsForHall
      parameters:
        - in: path
          name: hallId
          required: true
          schema:
            type: integer
          description: ID of the hall to retrieve movies
        - name: date
          in: query
          description: The date for which the movie sessions should be displayed in format YYYY-MM-DD
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CinemaSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []

    post:
      tags:
        - cinema sessions
      summary: Creates a new cinema session
      operationId: createCinemaSession
      parameters:
        - in: path
          name: hallId
          required: true
          schema:
            type: integer
          description: ID of the hall to create cinema session in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                movie_id:
                  type: integer
                  example: 1
                  description: ID of the movie being shown in the cinema session.
                start_time:
                  type: string
                  format: timestamp
                  description: Start time of the cinema session
                  example: 2024-05-18 20:00:00
                price:
                  type: number
                  format: float
                  example: 10.5
                  description: Price for the cinema session in GEL
      responses:
        '201':
          description: The newly created cinema session
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                    example: 1
                    description: Unique identifier of the cinema session. Generated on the server side.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []
  /cinema-sessions/{sessionId}:
    put:
      summary: Updates a specific cinema session
      operationId: updateCinemaSession
      tags:
        - cinema sessions
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
          description: ID of the cinema session to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CinemaSession'
      responses:
        '200':
          description: The cinema session was updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []

    delete:
      summary: Deletes a specific cinema session
      operationId: deleteCinemaSession
      tags:
        - cinema sessions
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
          description: ID of the cinema session to delete
      responses:
        '204':
          description: The cinema session was deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cinema-sessions/{sessionId}/seats:
    get:
      summary: Returns all available seats for the session
      operationId: getAvailableSeats
      tags:
        - cinema sessions
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
          description: ID of the cinema session to get seats available
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  example: [1, 2, 3, 4, 8, 9]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...